requirements.txt
telethon==1.29.1
python-dotenv==1.0.0